requirements.txt
SimpleITK=2.1.1.2
scikit-image==0.16.0
PyYAML=6.0
json=2.0.9
numpy==1.16.4
scipy==1.3.1
